Refactor Your Positions tabs
