+++ apps/hyperdrive-trading/src/ui/portfolio/PositionsSection/PositionsSection.tsx
@@ -5,13 +5,13 @@ import { Hyperdrive } from "src/appconfig/types";
 import { Well } from "src/ui/base/components/Well/Well";
 import { ClosedLongsTable } from "src/ui/portfolio/ClosedLongsTable/ClosedLongsTable";
 import { OpenLongsTable } from "src/ui/portfolio/OpenLongsTable/OpenLongsTable";
-import {
-  OpenOrClosedTab,
-  OpenOrClosedTabs,
-} from "src/ui/portfolio/OpenOrClosedTabs/OpenOrClosedTabs";
 import { OpenShortsTable } from "src/ui/portfolio/OpenShortsTable/OpenShortsTable";
+import {
+  PositionTab,
+  PositionTabs,
+} from "src/ui/portfolio/PositionTabs/PositionTabs";
 
-type PositionTab = "Longs" | "Shorts" | "LP";
+export type OpenOrClosedTab = "Open" | "Closed";
 
 export function PositionsSection({
   hyperdrive,
@@ -21,72 +21,69 @@ export function PositionsSection({
   // TODO: Move this state into router
   const [activePositionTab, setActivePositionTab] =
     useState<PositionTab>("Longs");
+
   // TODO: Move this state into router
   const [activeOpenOrClosedTab, setActiveOpenOrClosedTab] =
     useState<OpenOrClosedTab>("Open");
 
   return (
     <div>
-      <div className="mb-2 flex w-full items-center justify-between border-b border-neutral-content/30">
-        <span className="text-h5 font-thin text-neutral-content">
+      <div className="mb-4 flex w-full items-center justify-between border-b border-neutral-content/30">
+        <h2 className="text-h5 font-thin leading-h5 text-neutral-content">
           Your positions
-        </span>
-        <div className="daisy-tabs">
+        </h2>
+        <PositionTabs
+          setActivePositionTab={setActivePositionTab}
+          activePositionTab={activePositionTab}
+        />
+
+        <div className="daisy-tabs-boxed">
           <a
-            onClick={() => setActivePositionTab("Longs")}
-            className={classNames("daisy-tab-lifted daisy-tab daisy-tab-lg ", {
-              "daisy-tab-active glass": activePositionTab === "Longs",
+            onClick={() => setActiveOpenOrClosedTab("Open")}
+            className={classNames("daisy-tab", {
+              "daisy-tab-active !bg-base-300 !text-white":
+                activeOpenOrClosedTab === "Open",
             })}
           >
-            Longs
+            Open
           </a>
           <a
-            onClick={() => setActivePositionTab("Shorts")}
-            className={classNames("daisy-tab-lifted daisy-tab daisy-tab-lg", {
-              "daisy-tab-active glass": activePositionTab === "Shorts",
+            onClick={() => setActiveOpenOrClosedTab("Closed")}
+            className={classNames("daisy-tab", {
+              "daisy-tab-active !bg-base-300 !text-white":
+                activeOpenOrClosedTab === "Closed",
             })}
           >
-            Shorts
-          </a>
-          <a
-            onClick={() => setActivePositionTab("LP")}
-            className={classNames("daisy-tab-lifted daisy-tab daisy-tab-lg", {
-              "daisy-tab-active glass": activePositionTab === "LP",
-            })}
-          >
-            LP
+            Closed
           </a>
         </div>
       </div>
 
-      <div className="flex flex-col">
-        <OpenOrClosedTabs
-          setActiveOpenOrClosedTab={setActiveOpenOrClosedTab}
-          activeOpenOrClosedTab={activeOpenOrClosedTab}
-        />
-
-        {(() => {
-          switch (activePositionTab) {
-            case "Longs": {
-              if (activeOpenOrClosedTab === "Open") {
-                return <OpenLongsTable hyperdrive={hyperdrive} />;
+      <div className="flex w-full flex-col items-center">
+        <div className="w-full">
+          {(() => {
+            switch (activePositionTab) {
+              case "Longs": {
+                if (activeOpenOrClosedTab === "Open") {
+                  return <OpenLongsTable hyperdrive={hyperdrive} />;
+                }
+                return <ClosedLongsTable hyperdrive={hyperdrive} />;
               }
-              return <ClosedLongsTable hyperdrive={hyperdrive} />;
-            }
-            case "Shorts": {
-              if (activeOpenOrClosedTab === "Open") {
-                return <OpenShortsTable hyperdrive={hyperdrive} />;
+              case "Shorts": {
+                if (activeOpenOrClosedTab === "Open") {
+                  return <OpenShortsTable hyperdrive={hyperdrive} />;
+                }
+                // TODO: Wire this up
+                return <ClosedLongsTable hyperdrive={hyperdrive} />;
               }
-              // TODO: Wire this up
-              return <ClosedLongsTable hyperdrive={hyperdrive} />;
+              case "LP":
+                /* TODO: Wire this up */
+                return <Well>Under construction</Well>;
+              default:
+                assertNever(activePositionTab);
             }
-            case "LP":
-              /* TODO: Wire this up */
-              return <Well>Under construction</Well>;
-            default:
-              assertNever(activePositionTab);
-          }
-        })()}
+          })()}
+        </div>
       </div>
     </div>
   );
